==> scripts/100_postprocessing/01_genes_mappings.ipynb <==
%load_ext autoreload
%autoreload 2

import os
from glob import glob
import sqlite3

import numpy as np
import pandas as pd

import settings as conf

# Load all genes from models

gtex_models_files = glob(os.path.join(conf.GTEX_MODELS_DIR, '*.db'))
gtex_models_files[:3]

# should be 49 for GTEx v8 models
display(len(gtex_models_files))
assert len(gtex_models_files) == 49

# get gene id, gene name and gene type from sqlite files

all_models = []
all_models_size = []

for model_file in gtex_models_files:
    cnx = sqlite3.connect(model_file)
    df = pd.read_sql_query("SELECT gene, genename as gene_name, gene_type FROM extra", cnx)
    df = df.assign(gene_id=df['gene'].apply(lambda x: x.split('.')[0]))
    
    all_models_size.append(df.shape[0])
    all_models.append(df)

genes_mapping = pd.concat(all_models, ignore_index=True)

genes_mapping.shape

genes_mapping.head(3)

genes_mapping = genes_mapping.drop_duplicates()

genes_mapping.shape

assert genes_mapping.shape[0] == genes_mapping['gene'].unique().shape[0]

assert genes_mapping.shape[0] == genes_mapping['gene_id'].unique().shape[0]

genes_mapping.head()

## Save tissues

# save list of tissues
tissues_list = [os.path.basename(gtex_model).split('mashr_')[1].split('.db')[0] for gtex_model in gtex_models_files]

tissues_df = pd.DataFrame(tissues_list, columns=['tissue_name'])
tissues_df.index.rename('tissue_id', inplace=True)

tissues_df.shape

tissues_df.head()

os.makedirs(conf.DELIVERABLES_DIR, exist_ok=True)

tissues_file = os.path.join(conf.DELIVERABLES_DIR, 'tissues.tsv')
display(tissues_file)

tissues_df.to_csv(tissues_file, sep='\t')

def _get_col_stats(column_data):
    if column_data.dtype.name == 'object':
        return column_data.apply(lambda x: len(str(x))).max()
    else:
        return (column_data.min(), column_data.max())

    return None

{
    c : _get_col_stats(tissues_df[c])
    for c in tissues_df.columns
}

# Add gene bands

biomart_genes = pd.read_csv(conf.BIOMART_GENES_INFO_FILE, index_col='ensembl_gene_id')

def _get_gene_band(gene_id):
    if gene_id not in biomart_genes.index:
        return ''

    gene_data = biomart_genes.loc[gene_id]
    chrom = gene_data['chromosome_name']
    band = gene_data['band']

    return f'{chrom}{band}'

genes_mapping = genes_mapping.assign(band=genes_mapping['gene_id'].apply(lambda x: _get_gene_band(x)))

genes_mapping.head()

os.makedirs(conf.GENES_METADATA_DIR, exist_ok=True)

# for internal use
final_filename = os.path.join(conf.GENES_METADATA_DIR, 'genes_mappings.pkl')
display(final_filename)

genes_mapping.to_pickle(final_filename)

# for general use
os.makedirs(conf.DELIVERABLES_DIR, exist_ok=True)
final_filename = os.path.join(conf.DELIVERABLES_DIR, 'genes_mappings.tsv.gz')
display(final_filename)

genes_mapping.to_csv(final_filename, sep='\t', index=False)

# Save mappings

import pickle

### Gene id to name

gene_id_to_name = genes_mapping[['gene_id', 'gene_name']].set_index('gene_id').to_dict()['gene_name']
assert gene_id_to_name['ENSG00000000457'] == 'SCYL3'
assert gene_id_to_name['ENSG00000001036'] == 'FUCA2'

with open(os.path.join(conf.GENES_METADATA_DIR, 'genes_mapping_simplified-0.pkl'), 'wb') as f:
    pickle.dump(gene_id_to_name, f)

with open(os.path.join(conf.GENES_METADATA_DIR, 'genes_mapping_simplified-0.pkl'), 'rb') as f:
    gene_id_to_name = pickle.load(f)

assert gene_id_to_name['ENSG00000000457'] == 'SCYL3'
assert gene_id_to_name['ENSG00000001036'] == 'FUCA2'

### Gene id to band

gene_id_to_band = genes_mapping[['band', 'gene_id']].set_index('gene_id').to_dict()['band']
assert gene_id_to_band['ENSG00000000457'] == '1q24.2'
assert gene_id_to_band['ENSG00000000460'] == '1q24.2'
assert gene_id_to_band['ENSG00000001036'] == '6q24.2'

with open(os.path.join(conf.GENES_METADATA_DIR, 'genes_mapping_id_to_band.pkl'), 'wb') as f:
    pickle.dump(gene_id_to_band, f)

with open(os.path.join(conf.GENES_METADATA_DIR, 'genes_mapping_id_to_band.pkl'), 'rb') as f:
    gene_id_to_band = pickle.load(f)

assert gene_id_to_band['ENSG00000000457'] == '1q24.2'
assert gene_id_to_band['ENSG00000000460'] == '1q24.2'
assert gene_id_to_band['ENSG00000001036'] == '6q24.2'

### Gene name to id

gene_name_to_id = genes_mapping[['gene_id', 'gene_name']].set_index('gene_name').to_dict()['gene_id']
assert gene_name_to_id['SCYL3'] == 'ENSG00000000457'
assert gene_name_to_id['C1orf112'] == 'ENSG00000000460'

with open(os.path.join(conf.GENES_METADATA_DIR, 'genes_mapping_simplified-1.pkl'), 'wb') as f:
    pickle.dump(gene_name_to_id, f)

with open(os.path.join(conf.GENES_METADATA_DIR, 'genes_mapping_simplified-1.pkl'), 'rb') as f:
    gene_name_to_id = pickle.load(f)

assert gene_name_to_id['SCYL3'] == 'ENSG00000000457'
assert gene_name_to_id['C1orf112'] == 'ENSG00000000460'

### Gene name to band

gene_name_to_band = genes_mapping[['band', 'gene_name']].set_index('gene_name').to_dict()['band']
assert gene_name_to_band['SCYL3'] == '1q24.2'
assert gene_name_to_band['C1orf112'] == '1q24.2'
assert gene_name_to_band['FUCA2'] == '6q24.2'

with open(os.path.join(conf.GENES_METADATA_DIR, 'genes_mapping_name_to_band.pkl'), 'wb') as f:
    pickle.dump(gene_name_to_band, f)

with open(os.path.join(conf.GENES_METADATA_DIR, 'genes_mapping_name_to_band.pkl'), 'rb') as f:
    gene_name_to_band = pickle.load(f)

assert gene_name_to_band['SCYL3'] == '1q24.2'
assert gene_name_to_band['C1orf112'] == '1q24.2'
assert gene_name_to_band['FUCA2'] == '6q24.2'

# Deprecated

# Remove some genes with duplicated gene IDs

Keep only those IDs reported in GeneCards

# some gene names are duplicated (several gene ids per gene name)
genes_mapping.shape[0] == genes_mapping['gene_name'].unique().shape[0]

tmp = genes_mapping['gene_name'].value_counts()
tmp = tmp[tmp > 1]
display(tmp)

# taken from GeneCards
ens_ids_to_keep_map = {
    'LINC01115': 'ENSG00000237667',
    'MAL2': 'ENSG00000147676',
    'LYNX1': 'ENSG00000180155',
    'LINC01422': 'ENSG00000223704',
    'SPATA13': 'ENSG00000182957',
    'GOLGA8M': 'ENSG00000188626',
    'LINC01297': 'ENSG00000274827',
    'LINC00484': 'ENSG00000235641',
}

ens_ids_to_keep = list(ens_ids_to_keep_map.values())

genes_mapping[genes_mapping['gene_name'].isin(tmp.index)].sort_values('gene_name')

# keep only those from GeneCards
to_remove = genes_mapping[genes_mapping['gene_name'].isin(tmp.index) & (~genes_mapping['gene_id'].isin(ens_ids_to_keep))]
assert to_remove.shape[0] == len(ens_ids_to_keep)

to_remove

n_before = genes_mapping.shape[0]
display(n_before)
genes_mapping = genes_mapping.drop(to_remove.index)
display(genes_mapping.shape)
assert genes_mapping.shape[0] == (n_before - len(to_remove.index))
