==> scripts/100_postprocessing/05_clinvar.ipynb <==
%load_ext autoreload
%autoreload 2

import os
import pickle
from glob import glob
import re
from concurrent.futures import ProcessPoolExecutor, as_completed

import numpy as np
import pandas as pd
from scipy import stats
from sklearn.metrics import pairwise_distances

import settings as conf
from utils import is_number, chunker

# Load metadata

with open(os.path.join(conf.GENES_METADATA_DIR, 'genes_mapping_simplified-0.pkl'), 'rb') as f:
    genes_mapping_0 = pickle.load(f)

with open(os.path.join(conf.GENES_METADATA_DIR, 'genes_mapping_simplified-1.pkl'), 'rb') as f:
    genes_mapping_1 = pickle.load(f)

# ClinVar data loading

clinvar_data_file = os.path.join(conf.CLINVAR_DATA_FILE)
display(clinvar_data_file)
clinvar_data = pd.read_csv(clinvar_data_file, sep='\t')

display(clinvar_data.shape)
display(clinvar_data.head())

clinvar_data['SourceName'].value_counts().head()

# Prepare dataframe with gene associations

clinvar_non_empty_genes_data = clinvar_data.dropna(subset=['AssociatedGenes'])
display(clinvar_non_empty_genes_data.shape)

clinvar_unique_traits = pd.Index(clinvar_non_empty_genes_data['DiseaseName'].unique())
display(clinvar_unique_traits)

# clinvar unique genes
clinvar_unique_genes = pd.Index(clinvar_non_empty_genes_data['AssociatedGenes'].unique())
display(clinvar_unique_genes)

diseases_columns = {}

for disease_name, disease_data in clinvar_non_empty_genes_data.groupby('DiseaseName'):
    common_genes = disease_data['AssociatedGenes'].unique()
    diseases_columns[disease_name] = pd.Series(index=common_genes, data=1)

clinvar_genes_associations = pd.DataFrame(data=diseases_columns, index=clinvar_unique_genes, columns=clinvar_unique_traits).fillna(0).astype('uint8')

assert clinvar_genes_associations.shape == (clinvar_unique_genes.shape[0] , clinvar_unique_traits.shape[0])

clinvar_genes_associations.shape

clinvar_genes_associations.fillna(0).head()

assert clinvar_genes_associations.isna().sum().sum() == 0

# some testing
clinvar_genes_associations["Alzheimer's disease"].sort_values(ascending=False).head(10)

clinvar_genes_associations.sum().sort_values(ascending=False).head()

# define filename
os.makedirs(conf.GENE_ASSOC_DIR, exist_ok=True)
clinvar_genes_associations_filename = os.path.join(conf.GENE_ASSOC_DIR, 'clinvar-gene_associations.pkl.xz')
display(clinvar_genes_associations_filename)

clinvar_genes_associations.index.rename('gene', inplace=True)

assert clinvar_genes_associations.dtypes.value_counts().shape[0] == 1

clinvar_genes_associations.dtypes.value_counts()

clinvar_genes_associations.to_pickle(clinvar_genes_associations_filename)

# Export to CSV

os.makedirs(conf.GENE_ASSOC_DIR, exist_ok=True)

export_path = os.path.join(conf.GENE_ASSOC_DIR, 'clinvar-gene_associations.tsv.gz')
display(export_path)

clinvar_genes_associations.to_csv(export_path, sep='\t')

## Distribution of number of genes per trait

# import matplotlib.pyplot as plt
# import seaborn as sns

# clinvar_traits_n_genes = clinvar_genes_associations.sum()
# display(clinvar_traits_n_genes.shape)
# display(clinvar_traits_n_genes.sort_values(ascending=False).head())

# sns.countplot(clinvar_traits_n_genes[clinvar_traits_n_genes <= 10])
