==> scripts/100_postprocessing/05_fastenloc.ipynb <==
%load_ext autoreload
%autoreload 2

import os
from glob import glob
from concurrent.futures import ProcessPoolExecutor, as_completed

import numpy as np
import pandas as pd

import settings as conf
import metadata
from utils import is_number, chunker
from results.multixcan import MXPhenoInfo, MXPhenoResults

# fastENLOC reading functions

with open(conf.FASTENLOC_GTEX_TISSUES_FILE, 'r') as f:
    FASTENLOC_TISSUES_NAMES = set([x.strip() for x in f.readlines()])

# create summary of files
_path = os.path.join(conf.FASTENLOC_RESULTS_DIR['RapidGWASProject'], '**/*.enloc.sig.out')
display(_path)
all_fastenloc_results_files = glob(_path)

len(all_fastenloc_results_files)

PHENO_CODES = set(metadata.RAPID_GWAS_PHENO_INFO.index)

len(PHENO_CODES)

phenos = []
tissues = []

for f in all_fastenloc_results_files:
    f = os.path.basename(f)
    split = f.split('-')
    
    tiss = split[-1].split('.')[0]
    if tiss not in FASTENLOC_TISSUES_NAMES:
        tiss = '-'.join(split[-2:]).split('.')[0]
    
    phen = split[1]
    if phen not in PHENO_CODES:
        phen = '-'.join(split[1:3])
    
    phenos.append(phen)
    tissues.append(tiss)

results_summary = pd.DataFrame({'pheno': phenos, 'tissue': tissues, 'file': all_fastenloc_results_files})

display(results_summary.shape)
assert results_summary.shape[0] == int(conf.FASTENLOC_EXPECTED_PHENOTYPES['RapidGWASProject'] * conf.GTEX_MODELS_N_EXPECTED_TISSUES)

results_summary.head()

results_summary['pheno'].unique().shape

assert np.all([t in PHENO_CODES for t in results_summary['pheno'].unique()])

assert results_summary['tissue'].unique().shape[0] == 49

assert np.all([t in FASTENLOC_TISSUES_NAMES for t in results_summary['tissue'].unique()])

_all_tissues_in_results = results_summary['tissue'].unique()
assert np.all([t in _all_tissues_in_results for t in FASTENLOC_TISSUES_NAMES])

assert len(all_fastenloc_results_files) == int(conf.SMULTIXCAN_EXPECTED_PHENOTYPES['RapidGWASProject'] * conf.GTEX_MODELS_N_EXPECTED_TISSUES)

def read_fe(filename):
    try:
        fe_data = pd.read_csv(filename, sep='\s+', usecols=[0, 5], header=None).rename(columns={0: 'gene_cluster', 5: 'rcp'})
    except pd.errors.EmptyDataError as e:
        return 0.0
    fe_data = fe_data = fe_data.assign(gene=fe_data['gene_cluster'].apply(lambda x: x.split(':')[0]))
    fe_data = fe_data.assign(gene_cluster=fe_data['gene_cluster'].apply(lambda x: x.split(':')[1]))
    
    fe_max = fe_data.groupby('gene')['rcp'].sum()
    return fe_max

def read_pheno(pheno):
    _tmp = results_summary[results_summary['pheno'] == pheno]
    if _tmp.shape[0] == 0:
        return None
    
    _res = {}
    _all_genes = set()
    for idx, _data in _tmp.iterrows():
        tissue_data = read_fe(_data.file)
        _res[_data.tissue] = tissue_data
        if hasattr(tissue_data, 'index'):
            _all_genes.update(tissue_data.index)
    
    return pd.DataFrame(_res, index=_all_genes).fillna(-1).max(axis=1)

## Testing

### J15

# testing
t = read_fe(os.path.join(conf.FASTENLOC_RESULTS_DIR['RapidGWASProject'], 'J15/fastenloc-J15-Whole_Blood.enloc.sig.out'))

t.sort_values(ascending=False).head()

_gid = metadata.GENE_NAME_TO_ID_MAP['SLCO4C1']
display(_gid)
assert t[_gid].round(5) == 1.638e-02

# v1 results
#assert t[_gid] == (1.869e-07 + 1.578e-11 + 5.201e-12 + 1.409e-13 + 2.694e-11)

```parallel -j4 'cat {} | grep ENSG00000173930' ::: J15/fastenloc-J15-Whole_Blood.enloc.sig.out | column -t | sort -k6 -g```

_gid = metadata.GENE_NAME_TO_ID_MAP['ENPP4']
display(_gid)
assert t[_gid] == 0.002, t[_gid]

### 50_raw

# testing height
t = read_fe(os.path.join(conf.FASTENLOC_RESULTS_DIR['RapidGWASProject'], '50_raw/fastenloc-50_raw-Whole_Blood.enloc.sig.out'))

t.sort_values(ascending=False).head()

```parallel -j4 'cat {} | grep ENSG00000163703' ::: 50_raw/fastenloc-50_raw-Whole_Blood.enloc.sig.out | column -t | sort -k6 -g```

_gid = 'ENSG00000163703'
display(metadata.GENE_ID_TO_NAME_MAP[_gid])
assert t[_gid] == 9.999e-01

```parallel -j4 'cat {} | grep ENSG00000077984' ::: 50_raw/fastenloc-50_raw-Whole_Blood.enloc.sig.out | column -t | sort -k6 -g```

_gid = 'ENSG00000077984'
display(metadata.GENE_ID_TO_NAME_MAP[_gid])
_expected_value = (4.362e-04 + 5.395e-04 + 5.539e-04 + 5.756e-04 + 5.853e-04 + 5.877e-04 + 8.617e-04 + 1.150e-03)
display(_expected_value)
assert t[_gid].round(7) == _expected_value, t[_gid]

### J15: read complete phenotype

```parallel -j1 'grep -H ENSG00000116985 {}' ::: J15/fastenloc-J15-*.enloc.sig.out | column -t | sort -k6 -g```

t = read_pheno('J15')

t.sort_values(ascending=False).head()

t.loc['ENSG00000116985']

# v1: the sum is across clusters of Spleen (which has the maximum value)
#_expected_value = (1.124e-02 + 5.425e-07 + 1.153e-08 + 2.864e-06 + 1.644e-06)

# the maximum value is in Spleen (only one cluster)
_expected_value = 2.463e-02

display(_expected_value)
assert t.loc['ENSG00000116985'].round(5) == _expected_value

### 50_raw: read complete phenotype

t = read_pheno('50_raw')

t.sort_values(ascending=False).head()

t.loc['ENSG00000176593']

#### read all data and see which tissue has the max rcp

_all_files = glob(os.path.join(conf.FASTENLOC_RESULTS_DIR['RapidGWASProject'], '50_raw/fastenloc-50_raw-*.sig.out'))
_all_data = []

for f in _all_files:
    tmp = pd.read_csv(f, sep='\s+', header=None)
    _split = tmp[0].str.split(':', expand=True)
    tmp = tmp.assign(gene=_split[0])
    tmp = tmp.assign(gene_cluster=_split[1])
    tmp = tmp[tmp['gene'] == 'ENSG00000176593']
    tmp = tmp.assign(tissue=os.path.basename(f).split('fastenloc-')[1].split('.enloc.sig.out')[0])
    _all_data.append(tmp)

_all_data_df = pd.concat(_all_data, ignore_index=True)

_all_data_df.shape

_all_data_df.head()

_all_data_df.groupby('tissue')[5].sum().sort_values(ascending=False).head()

_all_data_df[_all_data_df['tissue'] == '50_raw-Esophagus_Muscularis']

parallel -j1 'grep -H ENSG00000176593 {}' ::: 50_raw/fastenloc-50_raw-*.enloc.sig.out | column -t | grep Esophagus_Muscularis

# the maximum value is in XXX
_expected_value = 1.280e-01 + 3.322e-03 + 7.676e-01 + 7.627e-01

display(_expected_value)
assert t.loc['ENSG00000176593'] == _expected_value

# ENLOC reading functions

display(conf.FASTENLOC_RESULTS_DIR['GTEX_GWAS'])

def read_enloc(enloc_result_path):
    enloc_filename = enloc_result_path.split('/')[-1]
    phenotype = enloc_filename.split('__PM__')[0]
    tissue = enloc_filename.split('__PM__')[1].split('.enloc.rst.gz')[0]
    
    enloc_result = pd.read_csv(enloc_result_path, sep='\s+')
    enloc_result = enloc_result.assign(gene_id=enloc_result['molecular_qtl_trait'].apply(lambda x: x.split('.')[0]))
    
    return enloc_result.groupby('gene_id')['locus_rcp'].max().rename(tissue)

def read_enloc_pheno(phenotype_code):
    all_tissues_data = {}
    all_genes = set()
    
    all_phenotype_files = glob(os.path.join(conf.FASTENLOC_RESULTS_DIR['GTEX_GWAS'], phenotype_code, f'{phenotype_code}__PM__*.enloc.rst'))
    
    for f in all_phenotype_files:
        f_data = read_enloc(f)
        all_tissues_data[f] = f_data
        all_genes.update(f_data.index)
    
    return pd.DataFrame(all_tissues_data, index=all_genes).fillna(-1).max(axis=1)

t = read_enloc_pheno('SSGAC_Education_Years_Pooled')

t.head()

# testing
assert t.loc['ENSG00000106113'] == 0.00
assert t.loc['ENSG00000164050'] == 0.006
assert t.loc['ENSG00000081377'] == 0.317

# Get Rapid GWAS phenotypes

all_smultixcan_files = glob(os.path.join(conf.SMULTIXCAN_RESULTS_DIR['RapidGWASProject'], '*.tsv.gz'))
all_smultixcan_phenotypes = [MXPhenoResults(p) for p in all_smultixcan_files]

assert len(all_smultixcan_files) == len(all_smultixcan_phenotypes) == conf.SMULTIXCAN_EXPECTED_PHENOTYPES['RapidGWASProject']

all_smultixcan_files[:5]

# Get GTEx GWAS phenotypes

_path = os.path.join(conf.SMULTIXCAN_RESULTS_DIR['GTEX_GWAS'], '*')
display(_path)
all_extra_results_files = glob(_path)
assert len(all_extra_results_files) == conf.SMULTIXCAN_EXPECTED_PHENOTYPES['GTEX_GWAS']

all_extra_results_files[:5]

_file_pattern = '(?P<code>[^/]+)_smultixcan_imputed_gwas_gtexv8mashr_ccn30\.txt'
all_extra_phenotypes = [MXPhenoResults(p, _file_pattern) for p in all_extra_results_files]
all_extra_phenotypes_plain_names = pd.Index([p.pheno_info.get_plain_name() for p in all_extra_phenotypes])

display(len(all_extra_phenotypes))
assert len(all_extra_phenotypes) == conf.SMULTIXCAN_EXPECTED_PHENOTYPES['GTEX_GWAS']

all_extra_phenotypes[0].pheno_info.get_plain_name()

all_extra_phenotypes_codes = [os.path.basename(f).split('_smultixcan_imputed_gwas_gtexv8mashr_ccn30')[0] for f in all_extra_results_files]
assert len(all_extra_phenotypes_codes) == conf.SMULTIXCAN_EXPECTED_PHENOTYPES['GTEX_GWAS']

all_extra_phenotypes_codes[:5]

all_extra_phenotypes_codes = set(all_extra_phenotypes_codes)

# Read all results

def read_generic(pheno_code):
    if pheno_code in all_extra_phenotypes_codes:
        return read_enloc_pheno(pheno_code)
    else:
        return read_pheno(pheno_code)

def _get_combined_results(phenos):
    return {pheno.pheno_info.get_plain_name() : read_generic(pheno.pheno_info.pheno_code) for pheno in phenos}

# testing
_tmp = _get_combined_results(all_smultixcan_phenotypes[:4])
assert len(_tmp) == 4

# testing
_tmp = _get_combined_results(all_extra_phenotypes[:4])
assert len(_tmp) == 4

_pending = read_generic('C_MYELOID-LEUKAEMIA')

_pending.head()

_pending = read_generic('SSGAC_Education_Years_Pooled')

_pending.head()

def _run_all(phenotype_chunks, n_jobs=conf.N_JOBS_HIGH):
    all_results = {}
    
    with ProcessPoolExecutor(max_workers=n_jobs) as executor:
        tasks = [executor.submit(_get_combined_results, chunk) for chunk in phenotype_chunks]
        for future in as_completed(tasks):
            res = future.result()
            all_results.update(res)

    return all_results

# phenotype_chunks = chunker(all_smultixcan_phenotypes[:5] + all_extra_phenotypes[:5], 2)
phenotype_chunks = chunker(all_smultixcan_phenotypes + all_extra_phenotypes, 200)

all_results = _run_all(phenotype_chunks)

## Save as DataFrame

assert len(all_results) == conf.FASTENLOC_EXPECTED_PHENOTYPES['RapidGWASProject'] + conf.FASTENLOC_EXPECTED_PHENOTYPES['GTEX_GWAS']

fastenloc_genes_associations = pd.DataFrame(all_results)
fastenloc_genes_associations.index.rename('gene_id', inplace=True)

assert fastenloc_genes_associations.index.is_unique

display(fastenloc_genes_associations.shape)
display(fastenloc_genes_associations.head())

# check genes id format
_tmp = pd.Series(fastenloc_genes_associations.index).apply(len).value_counts()
display(_tmp)
assert _tmp.shape[0] == 1
assert _tmp.index[0] == 15

# check columns data type
_tmp = fastenloc_genes_associations.dtypes.value_counts()
display(_tmp)
assert _tmp.shape[0] == 1

fastenloc_genes_associations.isna().sum().sum()

fastenloc_genes_associations.dropna(axis=0, how='all').shape

display(fastenloc_genes_associations.shape)

assert fastenloc_genes_associations.shape == (38062, conf.FASTENLOC_EXPECTED_PHENOTYPES['RapidGWASProject'] + conf.FASTENLOC_EXPECTED_PHENOTYPES['GTEX_GWAS'])

# some testing
t = fastenloc_genes_associations['SSGAC_Education_Years_Pooled']

t.head()

assert t.loc['ENSG00000106113'] == 0.00
assert t.loc['ENSG00000164050'] == 0.006
assert t.loc['ENSG00000081377'] == 0.317
assert pd.isnull(t.loc['ENSG00000000938'])

# some testing
t = fastenloc_genes_associations['50_raw-Standing_height']

t.head()

assert t.loc['ENSG00000176593'] == 1.280e-01 + 3.322e-03 + 7.676e-01 + 7.627e-01
#assert pd.isnull(t.loc['ENSG00000000938'])

### Save

os.makedirs(conf.GENE_ASSOC_DIR, exist_ok=True)

# Save
fastenloc_genes_associations_filename = os.path.join(conf.GENE_ASSOC_DIR, f'fastenloc-torus-rcp.pkl.xz')
display(fastenloc_genes_associations_filename)

fastenloc_genes_associations.to_pickle(fastenloc_genes_associations_filename)

### Save for publication

output_file = os.path.join(conf.GENE_ASSOC_DIR, 'fastenloc-torus-rcp.tsv.gz')
display(output_file)

fastenloc_genes_associations.to_csv(output_file, sep='\t', float_format='%.4e')

# test "for publication" file
_tmp = pd.read_csv(output_file, sep='\t', index_col='gene_id')

display(_tmp.shape)
assert _tmp.shape == fastenloc_genes_associations.shape

_tmp.head()

assert not _tmp.isin([np.inf, -np.inf]).any().any()

assert _tmp.isna().any().any()

_tmp_flat = pd.Series(_tmp.values.flatten()).dropna()
assert ((_tmp_flat >= 0) & (_tmp_flat <= 3)).all().all()

assert np.allclose(_tmp.values, fastenloc_genes_associations.values, atol=1e-320, rtol=1e-4, equal_nan=True)

### Save in HDF5 format for webapp

display(fastenloc_genes_associations_filename)

# FIXME: it's not necessary to load again
fastenloc_gene_associations = pd.read_pickle(fastenloc_genes_associations_filename)

fastenloc_gene_associations.shape

from utils import simplify_string_for_hdf5

FASTENLOC_HDF5_FILE = os.path.join(conf.GENE_ASSOC_DIR, 'fastenloc-torus-rcp.h5')
display(FASTENLOC_HDF5_FILE)

with pd.HDFStore(FASTENLOC_HDF5_FILE, mode='w', complevel=1) as store:
    for col in fastenloc_gene_associations.columns:
        print('.', flush=True, end='')
        
        clean_col = simplify_string_for_hdf5(col)
        store[clean_col] = fastenloc_gene_associations[col]

#### Testing

parallel -j1 'grep -H ENSG00000267462 {}' ::: 100001_raw/fastenloc-100001_raw-*.enloc.sig.out | column -t | sort -k6 -g

# testing
with pd.HDFStore(FASTENLOC_HDF5_FILE, mode='r') as store:
    store_keys = list(store.keys())
    assert len(store_keys) == fastenloc_gene_associations.shape[1]
    display(store_keys[:5])
    
    clean_col = simplify_string_for_hdf5('100001_raw-Food_weight')
    data = store[clean_col]
    assert data.shape == (38062,), data.shape
    assert data.loc['ENSG00000267462'].round(7) == 1.702e-04 + 1.803e-04, data.loc['ENSG00000267462']
