==> scripts/100_postprocessing/05_fastenloc_enrichment.ipynb <==
%load_ext autoreload
%autoreload 2

import os
from glob import glob
from concurrent.futures import ProcessPoolExecutor, as_completed

import numpy as np
import pandas as pd

import settings as conf
import metadata
from utils import is_number, chunker
from results.multixcan import MXPhenoInfo, MXPhenoResults

# fastENLOC reading functions

with open(conf.FASTENLOC_GTEX_TISSUES_FILE, 'r') as f:
    FASTENLOC_TISSUES_NAMES = set([x.strip() for x in f.readlines()])

# create summary of files
_path = os.path.join(conf.FASTENLOC_RESULTS_DIR['RapidGWASProject'], '**/*.enloc.enrich.out')
display(_path)
all_fastenloc_results_files = glob(_path)

len(all_fastenloc_results_files)

PHENO_CODES = set(metadata.RAPID_GWAS_PHENO_INFO.index)

len(PHENO_CODES)

phenos = []
tissues = []

for f in all_fastenloc_results_files:
    f = os.path.basename(f)
    split = f.split('-')
    
    tiss = split[-1].split('.')[0]
    if tiss not in FASTENLOC_TISSUES_NAMES:
        tiss = '-'.join(split[-2:]).split('.')[0]
    
    phen = split[1]
    if phen not in PHENO_CODES:
        phen = '-'.join(split[1:3])
    
    phenos.append(phen)
    tissues.append(tiss)

results_summary = pd.DataFrame({'pheno': phenos, 'tissue': tissues, 'file': all_fastenloc_results_files})

display(results_summary.shape)
assert results_summary.shape[0] == int(conf.FASTENLOC_EXPECTED_PHENOTYPES['RapidGWASProject'] * conf.GTEX_MODELS_N_EXPECTED_TISSUES)

results_summary.head()

results_summary['pheno'].unique().shape

assert np.all([t in PHENO_CODES for t in results_summary['pheno'].unique()])

assert results_summary['tissue'].unique().shape[0] == conf.GTEX_MODELS_N_EXPECTED_TISSUES

assert np.all([t in FASTENLOC_TISSUES_NAMES for t in results_summary['tissue'].unique()])

_all_tissues_in_results = results_summary['tissue'].unique()
assert np.all([t in _all_tissues_in_results for t in FASTENLOC_TISSUES_NAMES])

assert len(all_fastenloc_results_files) == int(conf.SMULTIXCAN_EXPECTED_PHENOTYPES['RapidGWASProject'] * conf.GTEX_MODELS_N_EXPECTED_TISSUES)

def read_fe_enrich(filename):
    enrich_data = pd.read_fwf(filename, header=None)
    enrich_data = enrich_data.rename(columns={0: 'type', 1: 'alpha', 2: 'se'})
    assert enrich_data.shape[0] == 3
    
    alpha0_data = enrich_data[enrich_data['type'] == 'Intercept']
    assert alpha0_data.shape[0] == 1
    alpha0_data = alpha0_data.iloc[0].loc[['alpha', 'se']]\
        .rename({'alpha': 'alpha0', 'se': 'alpha0_se'}).astype(float)
    
    alpha1_data = enrich_data[enrich_data['type'] == 'Enrichment (likelihood)']
    assert alpha1_data.shape[0] == 1
    alpha1_data = alpha1_data.iloc[0].loc[['alpha', 'se']]\
        .rename({'alpha': 'alpha1', 'se': 'alpha1_se'}).astype(float)
    
    alpha1_s_data = enrich_data[enrich_data['type'] == 'Enrichment (shrinkage)']
    assert alpha1_s_data.shape[0] == 1
    alpha1_s_data = alpha1_s_data.iloc[0].loc[['alpha', 'se']]\
        .rename({'alpha': 'alpha1_shrinkage', 'se': 'alpha1_shrinkage_se'}).astype(float)
    
    df = pd.concat([alpha0_data, alpha1_data, alpha1_s_data])
    return df
    

def read_fe_enrich_pheno(pheno): 
    _tmp = results_summary[results_summary['pheno'] == pheno] 
    if _tmp.shape[0] == 0: 
        return None 
     
    _res = {}
#     _all_genes = set()
    for idx, _data in _tmp.iterrows():
        tissue_data = read_fe_enrich(_data.file).rename(_data.pheno)
        _res[_data.tissue] = tissue_data
#         _all_genes.update(tissue_data.index) 
    
    df = pd.DataFrame.from_dict(_res, orient='index')#.T#, index=_all_genes)#.fillna(-1).max(axis=1)
    df.index.rename('tissue', inplace=True)
    df = df.reset_index().assign(trait=pheno)
    return df[['tissue', 'trait', 'alpha0', 'alpha0_se', 'alpha1', 'alpha1_se', 'alpha1_shrinkage', 'alpha1_shrinkage_se']]

results_summary.iloc[1].loc['file']

read_fe_enrich(results_summary.iloc[1].loc['file'])

read_fe_enrich_pheno('22617_9219').head()

### Testing

# testing
t = read_fe_enrich(os.path.join(conf.FASTENLOC_RESULTS_DIR['RapidGWASProject'], 'J15/fastenloc-J15-Whole_Blood.enloc.enrich.out'))

if 'alpha1_shrinkage' not in t.index:
    assert t['alpha0'] == -11.752
    assert t['alpha0_se'] == 0.350

    assert t['alpha1'] == -11.286
    assert t['alpha1_se'] == 764.307
else:
    assert t['alpha0'] == -11.620
    assert t['alpha0_se'] == 0.114

    assert t['alpha1'].round(3) == -13.851
    assert t['alpha1_se'] == 2491.946
    
    assert t['alpha1_shrinkage'] == -0.000
    assert t['alpha1_shrinkage_se'] == 1.000

t = read_fe_enrich_pheno('J15')

assert t.shape[0] == conf.GTEX_MODELS_N_EXPECTED_TISSUES

t.head()

if 'alpha1_shrinkage' not in t.columns:
    _t_tmp = t[t['tissue'] == 'Adipose_Visceral_Omentum'].iloc[0]
    assert _t_tmp['tissue'] == 'Adipose_Visceral_Omentum'
    assert _t_tmp['trait'] == 'J15'
    assert _t_tmp['alpha0'] == -11.747
    assert _t_tmp['alpha0_se'].round(3) == 0.349
    assert _t_tmp['alpha1'] == -11.061
    assert _t_tmp['alpha1_se'].round(3) == 877.891

    _t_tmp = t[t['tissue'] == 'Artery_Coronary'].iloc[0]
    assert _t_tmp['tissue'] == 'Artery_Coronary'
    assert _t_tmp['trait'] == 'J15'
    assert _t_tmp['alpha0'].round(3) == -11.736
    assert _t_tmp['alpha0_se'].round(3) == 0.348
    assert _t_tmp['alpha1'].round(3) == -11.979
    assert _t_tmp['alpha1_se'].round(3) == 1336.767
else:
    _t_tmp = t[t['tissue'] == 'Adipose_Visceral_Omentum'].iloc[0]
    assert _t_tmp['tissue'] == 'Adipose_Visceral_Omentum'
    assert _t_tmp['trait'] == 'J15'
    assert _t_tmp['alpha0'] == -11.618
    assert _t_tmp['alpha0_se'].round(3) == 0.114
    assert _t_tmp['alpha1'] == -14.123
    assert _t_tmp['alpha1_se'].round(3) == 2916.263
    assert _t_tmp['alpha1_shrinkage'] == -0.000
    assert _t_tmp['alpha1_shrinkage_se'].round(3) == 1.000

    _t_tmp = t[t['tissue'] == 'Artery_Coronary'].iloc[0]
    assert _t_tmp['tissue'] == 'Artery_Coronary'
    assert _t_tmp['trait'] == 'J15'
    assert _t_tmp['alpha0'].round(3) == -11.621
    assert _t_tmp['alpha0_se'].round(3) == 0.114
    assert _t_tmp['alpha1'].round(3) == -14.408
    assert _t_tmp['alpha1_se'].round(3) == 4208.734
    assert _t_tmp['alpha1_shrinkage'] == -0.000
    assert _t_tmp['alpha1_shrinkage_se'].round(3) == 1.000

# Get Rapid GWAS phenotypes

all_smultixcan_files = glob(os.path.join(conf.SMULTIXCAN_RESULTS_DIR['RapidGWASProject'], '*.tsv.gz'))
all_smultixcan_phenotypes = [MXPhenoResults(p) for p in all_smultixcan_files]

assert len(all_smultixcan_files) == len(all_smultixcan_phenotypes) == conf.SMULTIXCAN_EXPECTED_PHENOTYPES['RapidGWASProject']

all_smultixcan_files[:5]

# Read all results

def _get_combined_results(phenos):
    all_res = [read_fe_enrich_pheno(pheno.pheno_info.pheno_code) for pheno in phenos]
    return pd.concat(all_res, ignore_index=True)

# testing
_tmp = _get_combined_results(all_smultixcan_phenotypes[:4])
assert _tmp.shape[0] == int(4 * 49)

_pending = read_fe_enrich_pheno('C_MYELOID-LEUKAEMIA')

_pending.head()

def _run_all(phenotype_chunks, n_jobs=conf.N_JOBS_HIGH):
    all_results = []
    
    with ProcessPoolExecutor(max_workers=n_jobs) as executor:
        tasks = [executor.submit(_get_combined_results, chunk) for chunk in phenotype_chunks]
        for future in as_completed(tasks):
            res = future.result()
            all_results.append(res)

    return all_results

# phenotype_chunks = chunker(all_smultixcan_phenotypes[:5] + all_extra_phenotypes[:5], 2)
phenotype_chunks = chunker(all_smultixcan_phenotypes, 200)

all_results = _run_all(phenotype_chunks)

## Save as DataFrame

fastenloc_genes_associations = pd.concat(all_results, ignore_index=True)

display(fastenloc_genes_associations.shape)
display(fastenloc_genes_associations.head())

assert fastenloc_genes_associations.shape[0] == int(conf.FASTENLOC_EXPECTED_PHENOTYPES['RapidGWASProject'] * conf.GTEX_MODELS_N_EXPECTED_TISSUES)

assert fastenloc_genes_associations['tissue'].unique().shape[0] == conf.GTEX_MODELS_N_EXPECTED_TISSUES

assert fastenloc_genes_associations['trait'].unique().shape[0] == conf.FASTENLOC_EXPECTED_PHENOTYPES['RapidGWASProject']

assert fastenloc_genes_associations.shape == fastenloc_genes_associations.dropna().shape

# check columns data type
_tmp = fastenloc_genes_associations.dtypes.value_counts()
display(_tmp)
assert _tmp.shape[0] == 2

fastenloc_genes_associations.isna().sum().sum()

# some testing
t = fastenloc_genes_associations[fastenloc_genes_associations['trait'] == '22601_12253140']

t.head()

if 'alpha1_shrinkage' not in t.columns:
    tt = t[t['tissue'] == 'Brain_Frontal_Cortex_BA9'].iloc[0]
    assert tt.loc['alpha0'].round(3) == -10.781
    assert tt.loc['alpha0_se'].round(3) == 0.274
    assert tt.loc['alpha1'].round(3) == -12.584
    assert tt.loc['alpha1_se'] == 1050.572
else:
    tt = t[t['tissue'] == 'Brain_Frontal_Cortex_BA9'].iloc[0]
    assert tt.loc['alpha0'].round(3) == -10.731
    assert tt.loc['alpha0_se'].round(3) == 0.073
    assert tt.loc['alpha1'].round(3) == -14.893
    assert tt.loc['alpha1_se'] == 3230.940
    assert tt.loc['alpha1_shrinkage'].round(3) == -0.000
    assert tt.loc['alpha1_shrinkage_se'] == 1.000

if 'alpha1_shrinkage' not in t.columns:
    tt = t[t['tissue'] == 'Artery_Coronary'].iloc[0]
    assert tt.loc['alpha0'].round(3) == -10.783
    assert tt.loc['alpha0_se'].round(3) == 0.274
    assert tt.loc['alpha1'].round(3) == -12.448
    assert tt.loc['alpha1_se'].round(3) == 1042.841
else:
    tt = t[t['tissue'] == 'Artery_Coronary'].iloc[0]
    assert tt.loc['alpha0'].round(3) == -10.732
    assert tt.loc['alpha0_se'].round(3) == 0.073
    assert tt.loc['alpha1'].round(3) == -14.895
    assert tt.loc['alpha1_se'].round(3) == 3423.376
    assert tt.loc['alpha1_shrinkage'].round(3) == -0.000
    assert tt.loc['alpha1_shrinkage_se'] == 1.000

### Save

os.makedirs(conf.ANALYSES_DIR, exist_ok=True)

# Save
# new filename to save the alpha1 of newer fastENLOC runs
fastenloc_genes_associations_filename = os.path.join(conf.ANALYSES_DIR, f'fastenloc-alpha1-stats-v2.pkl.xz')

# filename for original alpha1 runs of fastENLOC
#fastenloc_genes_associations_filename = os.path.join(conf.ANALYSES_DIR, f'fastenloc-alpha1-stats.pkl.xz')

display(fastenloc_genes_associations_filename)

fastenloc_genes_associations.to_pickle(fastenloc_genes_associations_filename)

### Save in text format

# new filename to save the alpha1 of newer fastENLOC runs
output_file = os.path.join(conf.ANALYSES_DIR, 'fastenloc-alpha1-stats-v2.tsv.gz')

# filename for original alpha1 runs of fastENLOC
# output_file = os.path.join(conf.ANALYSES_DIR, 'fastenloc-alpha1-stats.tsv.gz')

display(output_file)

fastenloc_genes_associations.to_csv(output_file, index=False, sep='\t', float_format='%.3f')

# test "for publication" file
_tmp = pd.read_csv(output_file, sep='\t')

display(_tmp.shape)
assert _tmp.shape == fastenloc_genes_associations.shape

_tmp.head()

assert not _tmp.isin([np.inf, -np.inf]).any().any()

assert not _tmp.isna().any().any()

assert fastenloc_genes_associations.equals(_tmp)

Now a plot can be made with notebook `150_validation/35_fastenloc_alpha_se_plots.ipynb`
