==> scripts/100_postprocessing/05_spredixcan_direction_effect_consensus.ipynb <==
%load_ext autoreload
%autoreload 2

import os
import pickle
from glob import glob
from concurrent.futures import ProcessPoolExecutor, as_completed

import numpy as np
import pandas as pd
from scipy import stats

import settings as conf
from utils import is_number, chunker

# Load S-PrediXcan results

## From Rapid GWAS project

from results.spredixcan import PhenoResults

_path = os.path.join(conf.SPREDIXCAN_RESULTS_DIR['RapidGWASProject'] + '/*')
display(_path)
all_spredixcan_results_dirs = glob(_path)
display(len(all_spredixcan_results_dirs))
assert len(all_spredixcan_results_dirs) == conf.SPREDIXCAN_EXPECTED_PHENOTYPES['RapidGWASProject']

all_spredixcan_phenotypes = [PhenoResults(p) for p in all_spredixcan_results_dirs]

display(len(all_spredixcan_phenotypes))
assert len(all_spredixcan_phenotypes) == conf.SPREDIXCAN_EXPECTED_PHENOTYPES['RapidGWASProject']

## From GTEx GWAS manuscript

_path = os.path.join(conf.SPREDIXCAN_RESULTS_DIR['GTEX_GWAS'] + '/*')
display(_path)
all_extra_results_dirs = glob(_path)
display(len(all_extra_results_dirs))
assert len(all_extra_results_dirs) == conf.SPREDIXCAN_EXPECTED_PHENOTYPES['GTEX_GWAS']

all_extra_results_dirs[:5]

_file_pattern = 'spredixcan_igwas_gtexmashrv8_(?P<code>[^/]+)__PM__(?P<tissue>.+)\.csv$'
all_extra_phenotypes = [PhenoResults(p, _file_pattern) for p in all_extra_results_dirs]
all_extra_phenotypes_plain_names = pd.Index([p.pheno_info.get_plain_name() for p in all_extra_phenotypes])

display(len(all_extra_phenotypes))
assert len(all_extra_phenotypes) == conf.SMULTIXCAN_EXPECTED_PHENOTYPES['GTEX_GWAS']

# S-PrediXcan: direction of effect

## Effect direction: consensus

from results.spredixcan import PhenoResults

# from scipy import stats

# stats.norm.cdf(-1.96) * 2

# def _get_effect_direction(zscores):
#     pvalues = pd.Series(stats.norm.cdf(np.abs(zscores) * -1.0) * 2.0, index=zscores.index.tolist())
#     pvalues = pvalues[pvalues < pval_threshold]

#     if pvalues.shape[0] == 0:
#         return 0.0

#     zscores = zscores.loc[pvalues.index]
#     zscores_sign = np.sign(zscores)
#     return zscores_sign.value_counts().sort_values(ascending=False).index[0]

# _tmp = all_spredixcan_phenotypes[2]
# display(_tmp)

# data_dict = {t:_tmp.get_tissue_data(t, 'zscore', index_col='gene_simple') for t in _tmp.tissues}

# data = pd.DataFrame(data_dict)

# data.head()

# assert not data.isin([np.inf, -np.inf]).any().any()

# zscores = data.loc['ENSG00000137185']

# pval_threshold=1e-4

# pvalues = pd.Series(stats.norm.cdf(np.abs(zscores) * -1.0) * 2.0, index=zscores.index.tolist())

# pvalues = pvalues[pvalues < pval_threshold]

# pvalues

# zscores = zscores.loc[pvalues.index]

# zscores

# zscores_sign = np.sign(zscores)

# zscores_sign.value_counts().sort_values(ascending=False).index[0]

### Compute results

def _get_combined_results(phenos):
    return {
        pheno.pheno_info.get_plain_name():
            pheno.get_consensus_effect_direction()
        for pheno in phenos
    }

def _run_all(phenotype_chunks, n_jobs=conf.N_JOBS_HIGH):
    all_results = {}
    
    with ProcessPoolExecutor(max_workers=n_jobs) as executor:
        tasks = [executor.submit(_get_combined_results, chunk) for chunk in phenotype_chunks]
        for future in as_completed(tasks):
            res = future.result()
            all_results.update(res)
    
    return all_results

# phenotype_chunks = chunker(all_spredixcan_phenotypes[:5] + all_extra_phenotypes[:5], 2)
phenotype_chunks = chunker(all_spredixcan_phenotypes + all_extra_phenotypes, 25)

all_results = _run_all(phenotype_chunks, n_jobs=20)

len(all_results)

### Create DataFrame

_n_expected_phenos = np.sum(list(conf.SMULTIXCAN_EXPECTED_PHENOTYPES.values()))
display(_n_expected_phenos)
assert len(all_results) == _n_expected_phenos, len(all_results)

# the category dtype is for efficiency in storage/loading
spredixcan_genes_effect_directions = pd.DataFrame(all_results, dtype='category')
spredixcan_genes_effect_directions.index.rename('gene_name', inplace=True)

assert spredixcan_genes_effect_directions.index.is_unique

display(spredixcan_genes_effect_directions.shape)
display(spredixcan_genes_effect_directions.head())

# For this method we should not remove nans
# Remove genes with no results
#spredixcan_genes_effect_directions = spredixcan_genes_effect_directions.dropna(axis=0, how='all')

# how many entries are nan
spredixcan_genes_effect_directions.isna().sum().sum()

_tmp = pd.Series(spredixcan_genes_effect_directions.values.flatten()).dropna().astype(float).unique()
display(_tmp)
assert set(_tmp) == set([0, 1, -1]), set(_tmp)

display(f'Results shape: {spredixcan_genes_effect_directions.shape}')

assert spredixcan_genes_effect_directions.shape == (22518, _n_expected_phenos), spredixcan_genes_effect_directions.shape

## Testing

# _tmp.loc['CNCR_Insomnia_all'].sort_values(ascending=False).head()

# _tmp = all_spredixcan_phenotypes[2]

# _tmp.pheno_info.get_plain_name()

# data_dict = {t:_tmp.get_tissue_data(t, 'zscore', index_col='gene_simple') for t in _tmp.tissues}

# data = pd.DataFrame(data_dict)

# data.max(axis=1).sort_values()

spredixcan_genes_effect_directions.loc[
    [
        'ENSG00000137185',
        'ENSG00000000457',
        'ENSG00000095464',
        'ENSG00000228397',
        'ENSG00000279325',
        'ENSG00000000419',
    ],
    [
        '6157_3-Why_stopped_smoking_Health_precaution',
        'I50-Diagnoses_main_ICD10_I50_Heart_failure',
        'CNCR_Insomnia_all',
    ]
]

assert spredixcan_genes_effect_directions.loc['ENSG00000137185', '6157_3-Why_stopped_smoking_Health_precaution'] == -1.0
assert pd.isnull(spredixcan_genes_effect_directions.loc['ENSG00000000457', '6157_3-Why_stopped_smoking_Health_precaution'])
assert pd.isnull(spredixcan_genes_effect_directions.loc['ENSG00000095464', '6157_3-Why_stopped_smoking_Health_precaution'])

assert spredixcan_genes_effect_directions.loc['ENSG00000095464', 'I50-Diagnoses_main_ICD10_I50_Heart_failure'] == 1.0

assert pd.isnull(spredixcan_genes_effect_directions.loc['ENSG00000137185', 'CNCR_Insomnia_all'])
assert spredixcan_genes_effect_directions.loc['ENSG00000228397', 'CNCR_Insomnia_all'] == 1.0
assert spredixcan_genes_effect_directions.loc['ENSG00000279325', 'CNCR_Insomnia_all'] == -1.0

The code below was used to write the assert above; see for each gene if first and last (min and max) correspond to sign above

rapid_gwas_dir = conf.SPREDIXCAN_RESULTS_DIR['RapidGWASProject']
gtex_gwas_dir = conf.SPREDIXCAN_RESULTS_DIR['GTEX_GWAS']

%%bash -s "$rapid_gwas_dir"
cd $1/6157_3
parallel 'cat {} | cut -f1-3 -d, | column -t -s, | grep "ENSG00000137185"' ::: *.csv | sort -k3 -g
echo ""
parallel 'cat {} | cut -f1-3 -d, | column -t -s, | grep "ENSG00000000457"' ::: *.csv | sort -k3 -g | sed -e 1b -e '$!d'
echo ""
parallel 'cat {} | cut -f1-3 -d, | column -t -s, | grep "ENSG00000095464"' ::: *.csv | sort -k3 -g | sed -e 1b -e '$!d'

%%bash -s "$rapid_gwas_dir"
cd $1/I50
parallel 'cat {} | cut -f1-3 -d, | column -t -s, | grep "ENSG00000095464"' ::: *.csv | sort -k3 -g

%%bash -s "$gtex_gwas_dir"
cd $1/CNCR_Insomnia_all
parallel 'cat {} | cut -f1-3 -d, | column -t -s, | grep "ENSG00000137185"' ::: *.csv | sort -k3 -g # | sed -e 1b -e '$!d'
echo ""
parallel 'cat {} | cut -f1-3 -d, | column -t -s, | grep "ENSG00000228397"' ::: *.csv | sort -k3 -g # | sed -e 1b -e '$!d'
echo ""
parallel 'cat {} | cut -f1-3 -d, | column -t -s, | grep "ENSG00000279325"' ::: *.csv | sort -k3 -g # | sed -e 1b -e '$!d'

### Save

spredixcan_genes_effect_directions.shape

spredixcan_genes_effect_directions.head()

# Save
spredixcan_genes_effect_directions_filename = os.path.join(conf.GENE_ASSOC_DIR, f'spredixcan-mashr-effect_direction-consensus.pkl.xz')
display(spredixcan_genes_effect_directions_filename)

spredixcan_genes_effect_directions.to_pickle(spredixcan_genes_effect_directions_filename)

### Save in HDF5 format for webapp

spredixcan_genes_effect_directions = pd.read_pickle(spredixcan_genes_effect_directions_filename)

spredixcan_genes_effect_directions.shape

from utils import simplify_string_for_hdf5

os.makedirs(conf.GENE_ASSOC_DIR, exist_ok=True)

OUTPUT_HDF5_FILE = os.path.join(conf.GENE_ASSOC_DIR, 'spredixcan-mashr-effect_direction-consensus.h5')
display(OUTPUT_HDF5_FILE)

with pd.HDFStore(OUTPUT_HDF5_FILE, mode='w', complevel=1) as store:
    for col in spredixcan_genes_effect_directions.columns:
        #print('.', flush=True, end='')
        clean_col = simplify_string_for_hdf5(col)
        store[clean_col] = spredixcan_genes_effect_directions[col].astype(float)

# testing
with pd.HDFStore(OUTPUT_HDF5_FILE, mode='r') as store:
    store_keys = list(store.keys())
    assert len(store_keys) == spredixcan_genes_effect_directions.shape[1]
    display(store_keys[:5])
    
    clean_col = simplify_string_for_hdf5('6157_3-Why_stopped_smoking_Health_precaution')
    data = store[clean_col]
    assert data.shape == (22518,), data.shape
    assert data.loc['ENSG00000137185'] == -1.0
    assert pd.isnull(data.loc['ENSG00000000457'])
    assert pd.isnull(data.loc['ENSG00000095464'])
    
    clean_col = simplify_string_for_hdf5('I50-Diagnoses_main_ICD10_I50_Heart_failure')
    data = store[clean_col]
    assert data.shape == (22518,), data.shape
    assert data.loc['ENSG00000095464'] == 1.0
    
    clean_col = simplify_string_for_hdf5('CNCR_Insomnia_all')
    data = store[clean_col]
    assert data.shape == (22518,), data.shape
    assert pd.isnull(data.loc['ENSG00000137185'])
    assert data.loc['ENSG00000228397'] == 1.0
    assert data.loc['ENSG00000279325'] == -1.0
