==> scripts/100_postprocessing/06_smultixcan_and_fastenloc_merge.ipynb <==
%load_ext autoreload
%autoreload 2

import os
import pickle

import numpy as np
import pandas as pd

import settings as conf
import metadata

# Load S-MultiXcan results

smultixcan_pvalues_file = os.path.join(conf.GENE_ASSOC_DIR, f'smultixcan-mashr-pvalues.pkl.xz')
display(smultixcan_pvalues_file)

smultixcan_gene_associations = pd.read_pickle(smultixcan_pvalues_file)

smultixcan_gene_associations.shape

smultixcan_gene_associations.head(5)

# Load fastENLOC results

fastenloc_rcp_file = os.path.join(conf.GENE_ASSOC_DIR, f'fastenloc-torus-rcp.pkl.xz')
display(fastenloc_rcp_file)

fastenloc_gene_associations = pd.read_pickle(fastenloc_rcp_file)

fastenloc_gene_associations.shape

fastenloc_gene_associations.head(5)

# Common genes

common_genes = fastenloc_gene_associations.index.intersection(smultixcan_gene_associations.index)
display(common_genes)

# Match both matrices

multixcan_matrix_sorted = smultixcan_gene_associations.loc[common_genes]
display(multixcan_matrix_sorted.shape)
display(multixcan_matrix_sorted.head())

fastenloc_matrix_sorted = fastenloc_gene_associations.loc[common_genes, smultixcan_gene_associations.columns]
display(fastenloc_matrix_sorted.shape)
display(fastenloc_matrix_sorted.head())

# Merge all results

### S-MultiXcan

smultixcan_unstack = multixcan_matrix_sorted.unstack().rename('smultixcan_pvalue').reset_index()

smultixcan_unstack = smultixcan_unstack.rename(columns={
    'level_0': 'trait',
    'level_1': 'gene',
})

smultixcan_unstack = smultixcan_unstack.astype({
    'trait': 'category',
    'gene': 'category',
})

smultixcan_unstack.shape

smultixcan_unstack.dtypes

smultixcan_unstack.head()

del smultixcan_gene_associations, multixcan_matrix_sorted

### fastENLOC

fastenloc_unstack = fastenloc_matrix_sorted.unstack().rename('fastenloc_rcp').fillna(0.0).reset_index()

fastenloc_unstack = fastenloc_unstack.rename(columns={
    'level_0': 'trait',
    'level_1': 'gene',
})

fastenloc_unstack = fastenloc_unstack.astype({
    'trait': 'category',
    'gene': 'category',
})

fastenloc_unstack.shape

fastenloc_unstack.dtypes

fastenloc_unstack.head()

del fastenloc_gene_associations, fastenloc_matrix_sorted

### Merge

all_results = pd.merge(smultixcan_unstack, fastenloc_unstack, on=['trait', 'gene'], how='inner', validate='one_to_one')

all_results = all_results.assign(gene_name=all_results['gene'].apply(lambda x: metadata.GENE_ID_TO_NAME_MAP[x]))

all_results['gene_name'] = all_results['gene_name'].astype('category')

all_results = all_results[['trait', 'gene', 'gene_name', 'smultixcan_pvalue', 'fastenloc_rcp']]

all_results.shape

assert all_results.shape[0] == smultixcan_unstack.shape[0] == fastenloc_unstack.shape[0]

assert not all_results['trait'].isna().any()

assert not all_results['gene'].isna().any()

assert not all_results['gene_name'].isna().any()

assert not all_results['fastenloc_rcp'].isna().any()

# remove nan values from S-MultiXcan results
all_results = all_results.dropna(how='any')

all_results.shape

assert all_results.shape[0] == (91928861 - 963932)

all_results.head()

all_results.dtypes

# some testing
_pheno = '20096_1-Size_of_red_wine_glass_drunk_small_125ml'
_gene = 'ENSG00000000460'

_res = all_results[(all_results['trait'] == _pheno) & (all_results['gene'] == _gene)].iloc[0]

assert _res['smultixcan_pvalue'] == smultixcan_unstack[(smultixcan_unstack['trait'] == _pheno) & (smultixcan_unstack['gene'] == _gene)].iloc[0]['smultixcan_pvalue']
assert _res['fastenloc_rcp'] == fastenloc_unstack[(fastenloc_unstack['trait'] == _pheno) & (fastenloc_unstack['gene'] == _gene)].iloc[0]['fastenloc_rcp']

## Save

all_results_filename = os.path.join(conf.GENE_ASSOC_DIR, 'all_smultixcan_and_fastenloc_results.pkl.xz')
display(all_results_filename)

all_results.shape

all_results.head()

all_results.to_pickle(all_results_filename)

## Save for publication

all_results_filename = os.path.join(conf.GENE_ASSOC_DIR, 'all_smultixcan_and_fastenloc_results.tsv.gz')
display(all_results_filename)

all_results.to_csv(all_results_filename, sep='\t', float_format='%.4e', index=False)

_tmp = pd.read_csv(all_results_filename, sep='\t')

display(_tmp.shape)
assert _tmp.shape == all_results.shape

_tmp.head()

assert not _tmp.isin([np.inf, -np.inf]).any().any()

assert not _tmp.isna().any().any()

_tmp['smultixcan_pvalue'].isna().sum()

assert ((_tmp['smultixcan_pvalue'] >= 0) & (_tmp['smultixcan_pvalue'] <= 1)).all().all()

assert ((_tmp['fastenloc_rcp'] >= 0) & (_tmp['fastenloc_rcp'] <= 3)).all().all()

assert np.allclose(_tmp[['smultixcan_pvalue', 'fastenloc_rcp']].values, all_results[['smultixcan_pvalue', 'fastenloc_rcp']].values, atol=1e-15, rtol=1e-4)
