==> scripts/100_postprocessing/10_phenomexcan_and_clinvar-associations.ipynb <==
%load_ext autoreload
%autoreload 2

import os
#import pickle
from glob import glob
from concurrent.futures import ProcessPoolExecutor, as_completed

import numpy as np
import pandas as pd
#from scipy.cluster.hierarchy import linkage
from sklearn.metrics import pairwise_distances
#import matplotlib.pyplot as plt
#import seaborn as sns

import settings as conf
import metadata
from utils import chunker

# Load gene associations

## Clinvar

clinvar_genes_associations_filename = os.path.join(conf.GENE_ASSOC_DIR, 'clinvar-gene_associations.pkl.xz')
display(clinvar_genes_associations_filename)

clinvar_genes_associations = pd.read_pickle(clinvar_genes_associations_filename)

clinvar_genes_associations.shape

clinvar_genes_associations.head(5)

## S-MultiXcan

smultixcan_genes_associations_zscores_filename = os.path.join(conf.GENE_ASSOC_DIR, f'smultixcan-mashr-zscores.pkl.xz')
display(smultixcan_genes_associations_zscores_filename)

smultixcan_genes_associations_zscores = pd.read_pickle(smultixcan_genes_associations_zscores_filename)

smultixcan_genes_associations_zscores.shape

smultixcan_genes_associations_zscores.head(5)

### Compute unique list of genes and non-empty diseases

clinvar_genes_to_id = pd.Index([
    metadata.GENE_NAME_TO_ID_MAP[x]
    for x in clinvar_genes_associations.index
    if x in metadata.GENE_NAME_TO_ID_MAP
])
display(len(clinvar_genes_to_id))

shared_gene_ids = clinvar_genes_to_id.intersection(smultixcan_genes_associations_zscores.index)
display(len(shared_gene_ids))

shared_gene_names = pd.Index([metadata.GENE_ID_TO_NAME_MAP[x] for x in shared_gene_ids])
display(len(shared_gene_names))

assert len(shared_gene_ids) == len(shared_gene_names)

smultixcan_genes_associations_zscores = smultixcan_genes_associations_zscores.loc[shared_gene_ids]
clinvar_genes_associations = clinvar_genes_associations.loc[shared_gene_names]

_tmp = clinvar_genes_associations.sum()
_clinvar_diseases_to_remove = _tmp[_tmp == 0].index
display(_clinvar_diseases_to_remove.shape)

# remove diseases that have no-genes associated (after selecting just shared ones above)
clinvar_genes_associations = clinvar_genes_associations.drop(columns=_clinvar_diseases_to_remove)

_tmp = clinvar_genes_associations.sum()
assert _tmp[_tmp == 0].shape[0] == 0

display(smultixcan_genes_associations_zscores.shape)
display(clinvar_genes_associations.shape)

# Compute associations

smultixcan_squared_zscores = np.power(smultixcan_genes_associations_zscores.T.fillna(0).values, 2)

smultixcan_squared_zscores.shape

clinvar_genes_associations.shape

z2_sums = np.dot(smultixcan_squared_zscores, clinvar_genes_associations.values)

z2_sums.shape

n_clinvar_genes = clinvar_genes_associations.sum().to_frame().T
n_clinvar_genes = pd.concat([n_clinvar_genes] * z2_sums.shape[0], ignore_index=True).values

n_clinvar_genes.shape

pheno_corrs_results = z2_sums / n_clinvar_genes

pheno_corrs_results.shape

assoc_coef_df = pd.DataFrame(
    pheno_corrs_results,
    index=smultixcan_genes_associations_zscores.columns.copy(),
    columns=clinvar_genes_associations.columns.copy()
)

assert assoc_coef_df.shape == (smultixcan_genes_associations_zscores.shape[1], clinvar_genes_associations.shape[1])
display(assoc_coef_df.shape)

assoc_coef_df.head()

# testing
_pheno0 = "22617_5223-Job_SOC_coding_Metal_working_production_and_maintenance_fitters"
_pheno1 = "Alzheimer's disease"

_genes = clinvar_genes_associations[_pheno1].sort_values(ascending=False)
display(_genes.head(10))
_genes = _genes[_genes > 0].index
display(_genes)

_tmp = smultixcan_genes_associations_zscores[_pheno0].rename(index=metadata.GENE_ID_TO_NAME_MAP)
display(_tmp.head())

_tmp.loc[_genes]

_obs = np.power(_tmp.loc[_genes], 2).sum() / 6
display(_obs)
_exp = assoc_coef_df.loc[_pheno0, _pheno1]
display(_exp)
assert np.isclose(_obs, _exp), (_obs, _exp)

# testing
_pheno0 = '30040_raw-Mean_corpuscular_volume'
_pheno1 = "Alzheimer's disease"

_genes = clinvar_genes_associations[_pheno1].sort_values(ascending=False)
display(_genes.head(10))
_genes = _genes[_genes > 0].index
display(_genes)

_tmp = smultixcan_genes_associations_zscores[_pheno0].rename(index=metadata.GENE_ID_TO_NAME_MAP)
display(_tmp.head())

_tmp.loc[_genes]

_obs = np.power(_tmp.loc[_genes], 2).sum() / 6
display(_obs)
_exp = assoc_coef_df.loc[_pheno0, _pheno1]
display(_exp)
assert np.isclose(_obs, _exp), (_obs, _exp)

### Save

# sort index like smultixcan results
assoc_coef_df = assoc_coef_df.loc[smultixcan_genes_associations_zscores.columns]

assoc_coef_df.index.rename('phenomexcan_traits', inplace=True)

os.makedirs(conf.GENE_ASSOC_DIR, exist_ok=True)

assoc_coef_df_filename = os.path.join(conf.GENE_ASSOC_DIR, 'smultixcan_and_clinvar-z2.pkl.xz')
display(assoc_coef_df_filename)

assoc_coef_df.to_pickle(assoc_coef_df_filename)

pd.Series(assoc_coef_df.values.flatten()).describe()

assert not assoc_coef_df.isna().any().any()

assert not (assoc_coef_df < 0).any().any()

assert np.isfinite(assoc_coef_df).all().all()

# some testing
ukb_pheno = '20002_1111-Noncancer_illness_code_selfreported_asthma'
clinvar_pheno = 'Asthma, susceptibility to'

assert assoc_coef_df.loc[ukb_pheno, clinvar_pheno] == 28.748716710063185

### Save for publication

output_file = os.path.join(conf.GENE_ASSOC_DIR, 'smultixcan_and_clinvar-z2.tsv.gz')
display(output_file)

assoc_coef_df.to_csv(output_file, sep='\t', float_format='%.4e')

_tmp = pd.read_csv(output_file, sep='\t', index_col='phenomexcan_traits')

display(_tmp.shape)
assert _tmp.shape == assoc_coef_df.shape

_tmp.head()

assert not _tmp.isin([np.inf, -np.inf]).any().any()

assert not _tmp.isna().any().any()

assert ((_tmp >= 0) & (_tmp <= 1600)).all().all()

assert np.allclose(_tmp.values, assoc_coef_df.values, atol=1e-15, rtol=1e-4)
