==> scripts/100_postprocessing/20_traits_and_genes_info_files.ipynb <==
%load_ext autoreload
%autoreload 2

import os
import pickle

import numpy as np
import pandas as pd

import settings as conf
import metadata

## Load S-MultiXcan results

smultixcan_pvalues_file = os.path.join(conf.GENE_ASSOC_DIR, f'smultixcan-mashr-pvalues.pkl.xz')
display(smultixcan_pvalues_file)

smultixcan_gene_associations = pd.read_pickle(smultixcan_pvalues_file)

smultixcan_gene_associations.shape

smultixcan_gene_associations.head(5)

_tmp = smultixcan_gene_associations.unstack().dropna()

_tmp.shape

_tmp.head()

# Traits info file

metadata.RAPID_GWAS_PHENO_INFO.head()

metadata.GTEX_GWAS_PHENO_INFO.head()

from entity import Trait

pheno_codes = metadata.RAPID_GWAS_PHENO_INFO.index
pheno_codes = pheno_codes.union(metadata.GTEX_GWAS_PHENO_INFO.index)

assert len(pheno_codes) == len(metadata.RAPID_GWAS_PHENO_INFO.index) + len(metadata.GTEX_GWAS_PHENO_INFO.index)

pheno_codes

pheno_traits = [Trait(code) for code in pheno_codes]

assert len(pheno_traits) == len(pheno_codes)

pheno_traits[:5]

pheno_traits[-5:]

pheno_data_tuples = [
    (t.get_plain_name(), t.code, t.description, t.get_unique_description(), t.type, t.n, t.n_cases, t.n_controls, t.source)
    for t in pheno_traits
    if t.get_plain_name() in smultixcan_gene_associations.columns
]

pheno_info_df = pd.DataFrame.from_records(
    pheno_data_tuples,
#    index='full_code',
    columns='full_code short_code description unique_description type n n_cases n_controls source'.split()
)

pheno_info_df.index.rename('pheno_id', inplace=True)

pheno_info_df.head()

pheno_info_df.tail()

assert not pheno_info_df['description'].is_unique

assert pheno_info_df['unique_description'].is_unique

## Some stats

pheno_info_df.dtypes.loc['n'].name == 'object'

pheno_info_df['n_cases'].dtype

def _get_col_stats(column_data):
    if column_data.dtype.name == 'object':
        return column_data.apply(lambda x: len(str(x))).max()
    else:
        return (column_data.min(), column_data.max())

    return None

{
    c : _get_col_stats(pheno_info_df[c])
    for c in pheno_info_df.columns
}

## Save

os.makedirs(conf.DELIVERABLES_DIR, exist_ok=True)

pheno_info_df.to_csv(os.path.join(conf.DELIVERABLES_DIR, 'phenotypes_info.tsv.gz'), sep='\t')

# Genes info file

metadata.GENES_MAPPINGS.head()

genes_info = metadata.GENES_MAPPINGS.set_index('gene_id').loc[smultixcan_gene_associations.index.tolist()]

smultixcan_gene_associations.shape

genes_info = genes_info.reset_index()
genes_info.index.rename('gene_num_id', inplace=True)

assert genes_info.shape[0] == smultixcan_gene_associations.shape[0]

genes_info.head()

## Some stats

{
    c : _get_col_stats(genes_info[c])
    for c in genes_info.columns
}

## Save

genes_info.to_csv(os.path.join(conf.DELIVERABLES_DIR, 'genes_info.tsv.gz'), sep='\t')
